refactor: split conversation-store.ts into focused modules
